QR4File/requirements.txt
qrcode==7.4.2
python-resize-image==1.1.20
Pillow==9.0.0
tk==0.1.0
tk-tools==0.16.0
pyzbar==0.1.9
opencv-python==4.7.0.72
